faf_rpm: Do not use begin_nested() with autocommit
SQLAlchemy's SAVEPOINT transactions seem to cause unanticipated issues
when used in conjunction with autocommit mode. Use the explicit
`begin()` and `commit()` instead of `begin_nested()` and `flush()`.

This bug was introduced 088244e, but was missed in tests because they
are run with `autocommit=False` while production code is run with
`autocommit=True`.

Fixes #958